first pass at code refinement and exploration

+++ modules/simple_forms_api/app/services/simple_forms_api/share_point_service/service.rb
@@ -24,7 +24,7 @@ module SimpleFormsApi
       end
 
       ##
-      # Upload a PDF file of 5655 form to VHA SharePoint
+      # Upload a PDF file to VHA SharePoint
       #
       # @param form_contents [Hash] - The JSON of the form
       # @param form_submission [Form5655Submission] - Persisted submission of the form
@@ -33,23 +33,13 @@ module SimpleFormsApi
       # @return [Faraday::Response] - Response from SharePoint upload
       #
       def upload(form_contents:, form_submission:, station_id:)
-        @user = set_user_data(form_submission.user_account_id)
-        upload_response = upload_pdf(form_contents:, form_submission:,
-                                     station_id:)
+        set_user_data(form_submission.user_account_id)
+        upload_response = upload_pdf(form_contents:, form_submission:, station_id:)
+        list_item_id = fetch_list_item_id(upload_response)
 
-        list_item_id = get_pdf_list_item_id(upload_response)
-
-        resp = update_list_item_fields(list_item_id:, form_submission:, station_id:)
-        if resp.success?
-          StatsD.increment("#{STATSD_KEY_PREFIX}.success")
-        else
-          StatsD.increment("#{STATSD_KEY_PREFIX}.failure")
-        end
-        resp
+        update_sharepoint_item(list_item_id:, form_submission:, station_id:)
       rescue => e
-        StatsD.increment("#{STATSD_KEY_PREFIX}.failure")
-        Rails.logger.error('Sharepoint Upload failed', e.message)
-        raise e
+        handle_upload_error(e)
       end
 
       private
@@ -60,25 +50,35 @@ module SimpleFormsApi
       # @return [String] - The access token
       #
       def set_sharepoint_access_token
-        auth_response = auth_connection.post("/#{tenant_id}/tokens/OAuth/2", {
-                                               grant_type: 'client_credentials',
-                                               client_id: "#{client_id}@#{tenant_id}",
-                                               client_secret:,
-                                               resource: "#{resource}/#{sharepoint_url}@#{tenant_id}"
-                                             })
-
+        auth_response = auth_connection.post("/#{tenant_id}/tokens/OAuth/2", auth_params)
         auth_response.body['access_token']
+      end
+
+      def auth_params
+        {
+          grant_type: 'client_credentials',
+          client_id: "#{client_id}@#{tenant_id}",
+          client_secret:,
+          resource: "#{resource}/#{sharepoint_url}@#{tenant_id}"
+        }
       end
 
       def set_user_data(user_account_id)
         user_account = UserAccount.find(user_account_id)
-        user_profile = mpi_service.find_profile_by_identifier(identifier: user_account.icn,
-                                                              identifier_type: MPI::Constants::ICN)
+        user_profile = fetch_user_profile(user_account.icn)
+        @user = extract_user_info(user_profile)
+      end
 
+      # TODO: what is MPI service and can we use it?
+      def fetch_user_profile(icn)
+        mpi_service.find_profile_by_identifier(identifier: icn, identifier_type: MPI::Constants::ICN)
+      end
+
+      def extract_user_info(profile)
         {
-          ssn: user_profile.profile.ssn,
-          first_name: user_profile.profile.given_names.first,
-          last_name: user_profile.profile.family_name
+          ssn: profile.ssn,
+          first_name: profile.given_names.first,
+          last_name: profile.family_name
         }
       end
 
@@ -92,26 +92,34 @@ module SimpleFormsApi
       # @return [Faraday::Response]
       #
       def upload_pdf(form_contents:, form_submission:, station_id:)
-        pdf_path = PdfFill::Filler.fill_ancillary_form(form_contents, "#{form_submission.id}-#{station_id}", '5655')
-        fsr_pdf = File.open(pdf_path)
+        pdf_path = generate_pdf_path(form_contents, form_submission, station_id)
+        file_name = build_file_name(user)
 
-        file_name = "#{DateTime.now.strftime('%Y%m%dT%H%M%S')}_#{user[:ssn].last(4)}_#{user[:last_name].tr(' ', '_')}"
+        upload_to_sharepoint(pdf_path, file_name)
+      ensure
+        File.delete(pdf_path) if pdf_path
+      end
 
-        file_transfer_path =
-          "#{base_path}/_api/Web/GetFolderByServerRelativeUrl('#{base_path}/Submissions')" \
-          "/Files/add(url='#{file_name}.pdf',overwrite=true)"
+      # TODO: this needs to change
+      def generate_pdf_path(form_contents, form_submission, station_id)
+        PdfFill::Filler.fill_ancillary_form(form_contents, "#{form_submission.id}-#{station_id}", '5655')
+      end
 
+      def build_file_name(user)
+        "#{DateTime.now.strftime('%Y%m%dT%H%M%S')}_#{user[:ssn].last(4)}_#{user[:last_name].tr(' ', '_')}"
+      end
+
+      def upload_to_sharepoint(pdf_path, file_name)
         with_monitoring do
-          response = sharepoint_file_connection.post(file_transfer_path) do |req|
+          sharepoint_file_connection.post(file_transfer_url(file_name)) do |req|
             req.headers['Content-Type'] = 'octet/stream'
-            req.headers['Content-Length'] = fsr_pdf.size.to_s
-            req.body = Faraday::UploadIO.new(fsr_pdf, 'octet/stream')
+            req.body = Faraday::UploadIO.new(File.open(pdf_path), 'octet/stream')
           end
-
-          File.delete(pdf_path)
-
-          response
         end
+      end
+
+      def file_transfer_url(file_name)
+        "#{base_path}/_api/Web/GetFolderByServerRelativeUrl('#{base_path}/Submissions')/Files/add(url='#{file_name}.pdf',overwrite=true)"
       end
 
       ##
@@ -119,16 +127,22 @@ module SimpleFormsApi
       #
       # @param pdf_upload_response [Faraday::Response] - Network response from initial upload
       #
-      # @return [Number]
+      # @return [Integer]
       #
-      def get_pdf_list_item_id(pdf_upload_response)
-        uri = pdf_upload_response.body['d']['ListItemAllFields']['__deferred']['uri']
+      def fetch_list_item_id(pdf_upload_response)
+        list_item_uri = extract_list_item_uri(pdf_upload_response)
+        retrieve_list_item_id(list_item_uri)
+      end
+
+      def extract_list_item_uri(response)
+        response.body['d']['ListItemAllFields']['__deferred']['uri']
+      end
+
+      def retrieve_list_item_id(uri)
         path = uri.slice(uri.index(base_path)..-1)
-
         with_monitoring do
-          get_item_response = sharepoint_connection.get(path)
-
-          list_item_id = get_item_response.body.dig('d', 'ID')
+          response = sharepoint_connection.get(path)
+          list_item_id = response.body.dig('d', 'ID')
           raise ListItemNotFound if list_item_id.nil?
 
           list_item_id
@@ -136,69 +150,75 @@ module SimpleFormsApi
       end
 
       ##
-      # Populate data columns with properties needed by VHA
+      # Populate SharePoint list item fields with VHA data
       #
-      # @param list_item_id[Number] - ID of SharePoint list item
+      # @param list_item_id [Integer] - ID of SharePoint list item
       # @param form_submission [Form5655Submission] - Persisted form
       # @param station_id [String] - VHA Station identifier
       #
       # @return [Faraday::Response]
       #
-      def update_list_item_fields(list_item_id:, form_submission:, station_id:)
+      def update_sharepoint_item(list_item_id:, form_submission:, station_id:)
         path = "#{base_path}/_api/Web/Lists/GetByTitle('Submissions')/items(#{list_item_id})"
         with_monitoring do
           sharepoint_connection.post(path) do |req|
             req.headers['Content-Type'] = 'application/json;odata=verbose'
             req.headers['X-HTTP-METHOD'] = 'MERGE'
             req.headers['If-Match'] = '*'
-            req.body = {
-              '__metadata' => {
-                'type' => 'SP.Data.SubmissionsItem'
-              },
-              'StationId' => station_id,
-              'UID' => form_submission.id,
-              'SSN' => user[:ssn],
-              'Name1' => "#{user[:last_name]}, #{user[:first_name]}"
-            }.to_json
+            req.body = build_item_payload(form_submission, station_id).to_json
           end
         end
       end
 
+      def build_item_payload(form_submission, station_id)
+        {
+          '__metadata' => { 'type' => 'SP.Data.SubmissionsItem' },
+          'StationId' => station_id,
+          'UID' => form_submission.id,
+          'SSN' => user[:ssn],
+          'Name1' => "#{user[:last_name]}, #{user[:first_name]}"
+        }
+      end
+
+      def handle_upload_error(error)
+        StatsD.increment("#{STATSD_KEY_PREFIX}.failure")
+        Rails.logger.error('SharePoint upload failed', error.message)
+        raise error
+      end
+
       def auth_connection
-        Faraday.new(url: authentication_url, headers: auth_headers) do |conn|
-          conn.request :url_encoded
-          conn.use :breakers
-          conn.use Faraday::Response::RaiseError
-          conn.response :raise_custom_error, error_prefix: service_name
-          conn.response :json
-          conn.response :betamocks if mock_enabled?
-          conn.adapter Faraday.default_adapter
+        @auth_connection ||= Faraday.new(url: authentication_url, headers: auth_headers) do |conn|
+          configure_connection(conn)
         end
       end
 
       def sharepoint_connection
-        Faraday.new(url: "https://#{sharepoint_url}", headers: sharepoint_headers) do |conn|
-          conn.request :json
-          conn.use :breakers
-          conn.use Faraday::Response::RaiseError
-          conn.response :raise_custom_error, error_prefix: service_name
-          conn.response :json
-          conn.response :betamocks if mock_enabled?
-          conn.adapter Faraday.default_adapter
+        @sharepoint_connection ||= Faraday.new(url: "https://#{sharepoint_url}", headers: sharepoint_headers) do |conn|
+          configure_connection(conn)
         end
       end
 
       def sharepoint_file_connection
-        Faraday.new(url: "https://#{sharepoint_url}", headers: sharepoint_headers) do |conn|
-          conn.request :multipart
-          conn.request :url_encoded
-          conn.use :breakers
-          conn.use Faraday::Response::RaiseError
-          conn.response :raise_custom_error, error_prefix: service_name
-          conn.response :json
-          conn.response :betamocks if mock_enabled?
-          conn.adapter Faraday.default_adapter
+        @sharepoint_file_connection ||= Faraday.new(url: "https://#{sharepoint_url}",
+                                                    headers: sharepoint_headers) do |conn|
+          configure_file_connection(conn)
         end
+      end
+
+      def configure_connection(conn)
+        conn.request :json
+        conn.use :breakers
+        conn.use Faraday::Response::RaiseError
+        conn.response :raise_custom_error, error_prefix: service_name
+        conn.response :json
+        conn.response :betamocks if mock_enabled?
+        conn.adapter Faraday.default_adapter
+      end
+
+      def configure_file_connection(conn)
+        conn.request :multipart
+        conn.request :url_encoded
+        configure_connection(conn)
       end
 
       ##
@@ -207,9 +227,7 @@ module SimpleFormsApi
       # @return [Hash]
       #
       def auth_headers
-        {
-          'Content-Type' => 'application/x-www-form-urlencoded'
-        }
+        { 'Content-Type' => 'application/x-www-form-urlencoded' }
       end
 
       ##
@@ -224,12 +242,13 @@ module SimpleFormsApi
         }
       end
 
+      # TODO: we need to set this to VBA
       def initialize_settings
-        @settings = Settings.vha.sharepoint
+        Settings.vha.sharepoint
       end
 
       def mpi_service
-        @service ||= MPI::Service.new
+        @mpi_service ||= MPI::Service.new
       end
 
       ##
